# Conflicts:
#	native/cocos/audio/android/PcmAudioService.h
#	native/cocos/audio/android/audio.h
#	native/cocos/audio/android/utils/Utils.cpp
#	native/cocos/audio/common/utils/include/primitives.h
#	native/cocos/audio/common/utils/include/tinysndfile.h
#	native/cocos/base/TypeDef.h
#	native/cocos/bindings/auto/jsb_cocos_auto.cpp
#	native/cocos/bindings/auto/jsb_cocos_auto.h
#	native/cocos/bindings/manual/jsb_module_register.cpp
#	native/cocos/engine/Engine.cpp
#	native/external-config.json
#	pal/env/native/env.ts
#	pal/screen-adapter/native/screen-adapter.ts
#	platforms/native/builtin/index.js
#	platforms/native/engine/jsb-gfx.js
#	platforms/native/engine/jsb-loader.js
#	scripts/build-engine/package-lock.json
#	scripts/build-engine/package.json
#	scripts/build-engine/src/index.ts

+++ native/cocos/network/HttpRequest.h
@@ -241,8 +241,8 @@ public:
         _timeoutInSeconds = timeoutInSeconds;
     }
 
-    inline float getTimeout() const {
-        return _timeoutInSeconds;
+    inline long getTimeout() const {
+        return static_cast<long>(_timeoutInSeconds);
     }
 
 protected:
